organized components into subdirectories based on what top level component they are for
